mobile footer (wip)
