+++ src/app/portal/common/common.module.ts
@@ -1,8 +1,8 @@
-import { MatButtonModule } from '@angular/material/button';
-import { PortalFooterComponent } from './components/footer/portal-footer.component';
 /* eslint-disable  @typescript-eslint/no-explicit-any */
 import { CommonModule } from '@angular/common';
 import { NgModule, Type } from '@angular/core';
+import { MatButtonModule } from '@angular/material/button';
+import { MatExpansionModule } from '@angular/material/expansion';
 import { MatMenuModule } from '@angular/material/menu';
 import { MatToolbarModule } from '@angular/material/toolbar';
 import { RouterModule } from '@angular/router';
@@ -10,6 +10,9 @@ import { FontAwesomeModule } from '@fortawesome/angular-fontawesome';
 import { EffectsModule } from '@ngrx/effects';
 import { StoreModule } from '@ngrx/store';
 import { CoreModule } from 'src/app/core/core.module';
+import { PortalFooterDesktopComponent } from './components/footer/desktop/portal-footer-desktop.component';
+import { PortalFooterMobileComponent } from './components/footer/mobile/portal-footer-mobile.component';
+import { PortalFooterComponent } from './components/footer/portal-footer.component';
 import { PortalHeaderComponent } from './components/header/portal-header.component';
 import { PortalMenuComponent } from './components/menu/portal-menu.component';
 import { commonFeatureKey } from './constants/common.constants';
@@ -19,6 +22,8 @@ import { commonReducer } from './state/common.reducer';
 
 const components: Type<any>[] = [
   PortalFooterComponent,
+  PortalFooterDesktopComponent,
+  PortalFooterMobileComponent,
   PortalHeaderComponent,
   PortalMenuComponent,
 ];
@@ -35,6 +40,7 @@ const framework: Type<any>[] = [
 
 const materials: Type<any>[] = [
   FontAwesomeModule,
+  MatExpansionModule,
   MatButtonModule,
   MatToolbarModule,
 ];
