+++ src/app/portal/common/components/footer/desktop/portal-footer-desktop.component.scss
@@ -1,0 +1,29 @@
+footer {
+  display: grid;
+  grid-template-columns: repeat(3, 1fr) min-content;  
+
+  padding: var(--gap);
+
+  color: var(--color-primary-900);
+  background-color: white;
+  border-top: 1px solid var(--color-border);
+}
+
+nav {
+  display: flex;
+  flex-flow: column;
+  gap: 1rem;
+}
+
+a {
+  color: var(--color-primary-900);
+}
+
+aside {
+  place-self: end;
+
+  nav:last-child {
+    flex-flow: row;
+    justify-content: center;
+  }
+}
