+++ src/app/portal/common/components/footer/mobile/portal-footer-mobile.component.html
@@ -1,0 +1,44 @@
+<footer>
+  <mat-accordion multi>
+    <mat-expansion-panel *ngFor="let menuItem of menu | async">
+      <mat-expansion-panel-header>
+        <mat-panel-title>
+          <h3>{{menuItem | translatable: 'name' | async}}</h3>
+        </mat-panel-title>
+      </mat-expansion-panel-header>
+      <a *ngFor="let subItem of menuItem.subMenuItems" (click)="route(subItem)">
+        {{subItem | translatable: 'name' | async}}
+      </a>
+    </mat-expansion-panel>
+  </mat-accordion>
+  <aside>
+    <nav>
+      <a *ngFor="let app of apps | async"
+        [href]="app.url"
+        [title]="app.platform?.name">
+        <app-image
+          [src]="app.platform?.key === 'google'
+            ? 'assets/google-play.svg'
+            : 'assets/app-store.svg'"></app-image>
+      </a>
+    </nav>
+
+    <nav>
+      <a *ngFor="let item of socialMedia | async"
+        mat-icon-button 
+        [href]="item.url"
+        [title]="item.name">
+        <fa-icon
+          size="2x" 
+          [icon]="['fab', icon(item.icon)]">
+        </fa-icon>
+      </a>
+      <button mat-icon-button>
+        <fa-icon
+          size="2x" 
+          [icon]="['fas', 'rss']">
+        </fa-icon>
+      </button>
+    </nav>
+  </aside>
+</footer>
