+++ src/app/portal/common/components/footer/mobile/portal-footer-mobile.component.scss
@@ -1,0 +1,9 @@
+footer {
+  color: var(--color-primary-900);
+  background-color: white;
+  border-top: 1px solid var(--color-border);
+}
+
+mat-expansion-panel {
+  box-shadow: none !important;
+}
