+++ src/app/portal/common/components/footer/portal-footer.component.html
@@ -1,38 +1,2 @@
-<footer>
-  <nav *ngFor="let menuItem of menu | async">
-    <h3>{{menuItem | translatable: 'name' | async}}</h3>
-    <a *ngFor="let subItem of menuItem.subMenuItems" (click)="route(subItem)">
-      {{subItem | translatable: 'name' | async}}
-    </a>
-  </nav>
-  <aside>
-    <nav>
-      <a *ngFor="let app of apps | async"
-        [href]="app.url"
-        [title]="app.platform?.name">
-        <app-image
-          [src]="app.platform?.key === 'google'
-            ? 'assets/google-play.svg'
-            : 'assets/app-store.svg'"></app-image>
-      </a>
-    </nav>
-
-    <nav>
-      <a *ngFor="let item of socialMedia | async"
-        mat-icon-button 
-        [href]="item.url"
-        [title]="item.name">
-        <fa-icon
-          size="2x" 
-          [icon]="['fab', icon(item.icon)]">
-        </fa-icon>
-      </a>
-      <button mat-icon-button>
-        <fa-icon
-          size="2x" 
-          [icon]="['fas', 'rss']">
-        </fa-icon>
-      </button>
-    </nav>
-  </aside>
-</footer>
+<app-portal-footer-desktop></app-portal-footer-desktop>
+<app-portal-footer-mobile></app-portal-footer-mobile>
