+++ src/app/portal/common/components/footer/portal-footer.component.scss
@@ -1,30 +1,13 @@
-footer {
-  display: grid;
-  grid-template-columns: repeat(3, 1fr) min-content;  
+@import 'src/styles/mixins.scss';
 
-  padding: var(--gap);
-
-  color: var(--color-primary-900);
-  background-color: white;
-  border-top: 1px solid var(--color-border);
+@include responsive(tablet) {
+  app-portal-footer-desktop {
+    display: none;
+  }
 }
 
-nav {
-  display: flex;
-  flex-flow: column;
-  gap: 1rem;
-}
-
-a {
-  color: var(--color-primary-900);
-}
-
-aside {
-  width: 10rem;
-  place-self: end;
-
-  nav:last-child {
-    flex-flow: row;
-    justify-content: center;
+@include responsive(desktop) {
+  app-portal-footer-mobile {
+    display: none;
   }
 }
